refactor: one file suffice
